requirements.txt
flask
flask-classful
pyyaml
requests
